requirements.txt
pandas>=1.5.0
pymongo>=4.0.0
pymysql>=1.0.0
requests>=2.28.0
lxml>=4.9.0
jpholiday>=0.1.8
pytz>=2022.1
selenium>=4.0.0
webdriver-manager>=3.8.0
python-dotenv>=0.19.0
paramiko>=2.9.0
pyperclip>=1.8.0